Refactored the teleop package and added depth and yaw control to ps3_control
