+++ riptide_teleop/CMakeLists.txt
@@ -16,9 +16,5 @@ roslint_cpp()
 
 include_directories(include ${catkin_INCLUDE_DIRS})
 
-add_executable(accel_cmdr src/accel_cmd.cpp)
-target_link_libraries(accel_cmdr ${catkin_LIBRARIES})
-
-add_executable(commander src/commander.cpp)
-target_link_libraries(commander ${catkin_LIBRARIES})
-add_dependencies(commander riptide_msgs_gencpp)
+add_executable(ps3_control src/ps3_control.cpp)
+target_link_libraries(ps3_control ${catkin_LIBRARIES})
