--- riptide_teleop/launch/commander.launch
@@ -1,4 +1,0 @@
-<launch>
-  <node pkg="joy" type="joy_node" name="joystick_driver" />
-  <node pkg="riptide_teleop" type="commander" name="commander" />
-</launch>
